+++ DWSIM.Interfaces/IGraphicObject.vb
@@ -57,9 +57,9 @@ Imports DWSIM.Interfaces.Enums.GraphicObjects
 
     Property IsConnector() As Boolean
 
-    Property X() As Integer
+    Property X() As Single
 
-    Property Y() As Integer
+    Property Y() As Single
 
     Property Name() As String
